+++ Revit2018_Engine/Convert/ToSI.cs
@@ -20,30 +20,30 @@ namespace BH.Engine.Revit
     /// </summary>
     public static partial class Convert
     {
-        public static double FromSI(this double Value, UnitType UnitType)
+        public static double ToSI(this double Value, UnitType UnitType)
         {
             switch(UnitType)
             {
                 case UnitType.UT_Length:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_METERS);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_METERS);
                 case UnitType.UT_Mass:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_KILOGRAMS_MASS);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_KILOGRAMS_MASS);
                 case UnitType.UT_Electrical_Current:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_AMPERES);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_AMPERES);
                 case UnitType.UT_HVAC_Temperature:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_KELVIN);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_KELVIN);
                 case UnitType.UT_Weight:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_KILOGRAMS_MASS);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_KILOGRAMS_MASS);
                 case UnitType.UT_HVAC_Pressure:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_PASCALS);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_PASCALS);
                 case UnitType.UT_Piping_Pressure:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_PASCALS);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_PASCALS);
                 case UnitType.UT_HVAC_Velocity:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_METERS_PER_SECOND);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_METERS_PER_SECOND);
                 case UnitType.UT_Area:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_SQUARE_METERS);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_SQUARE_METERS);
                 case UnitType.UT_Volume:
-                    return UnitUtils.ConvertToInternalUnits(Value, DisplayUnitType.DUT_CUBIC_METERS);
+                    return UnitUtils.ConvertFromInternalUnits(Value, DisplayUnitType.DUT_CUBIC_METERS);
                 default:
                     return Value;
             }
