+++ Revit2018_Engine/Revit2018_Engine.csproj
@@ -75,8 +75,8 @@
     <Compile Include="Convert\Enums.cs" />
     <Compile Include="Convert\ToBHoM.cs" />
     <Compile Include="Convert\ToRevit.cs" />
-    <Compile Include="Convert\FromSI .cs" />
     <Compile Include="Convert\ToSI.cs" />
+    <Compile Include="Convert\FromSI.cs" />
     <Compile Include="Modify\AssignAdjacentSpaces.cs" />
     <Compile Include="Modify\RemoveIdentifiers.cs" />
     <Compile Include="Modify\SetCustomData.cs" />
